Add files via upload
Search button in RecipeSearch fragment includes same functionality as green plus button in addition to performing the recipe search. A single ingredient typed  or the last ingredient typed in will be added to the ingredient search list upon clicking the search button.

Upon removing ingredients from the search list any that came from the mSuggestedIngredientList are restore in alphabetical order, but user typed ingredients are discarded so as not to allow typos and nonsense to be added to the mSuggestedIngredientList.